requirements.txt

streamlit
fpdf2
pandas
supabase
python-dotenv
requests
